Refactor object property storage and management
Currently the properties are stored in linked list and in pairs
because of the memory alignment. This patch introduces a new
array based storage for the properties that helps to eliminate
tha property pair conception. In this case a property can be
used as a single unit.

JerryScript-DCO-1.0-Signed-off-by: Roland Takacs rtakacs@inf.u-szeged.hu